fixed the duplicate-except-for-weight bug
This was limitation #1 in the commit message for defefc8.

There is another very serious bug causing wrong results, where
some intervals' weights are double counted (or more), or some such
effect. I haven't figured that our or fixed it. Right now this code
is wrong. I don't think that was introduced in this commit.