basic idea for web version, no visualization yet
(1) Probably many bugs.

(2) The solution should list each interval including the weight,
    but currently the weights are not listed.

(3) Not adequately self-documenting. The UI should show some
    indication of what the input format is/means.

(4) The total weight should be printed after "Pyodide loaded
    successfully!" on the initial run.

(5) Needs visualization. That's the main point of the web version.

(6) Some of the code in wi.py, in perhaps a later and/or simplified
    form, will need to be backported to main/no-web, for when the
    command-line version is made. (The web branch, and main branch
    when it is merged, should contain that version and the web
    version. The no-web branch should contain just that version.)