finished BUGS.md (I think); more work on ALGORITHM.md
I mean I think I documented all the bugs I had been thinking of and
intending to document, not that there are no more bugs in the
program.

One of the changes I made in ALGORITHM.md is to remove the summary
near the top of the relationship between my algorithm and that of
Kleinberg & Tardos. I don't think that summary was really correct.
Their algorithm does not simply check the two subproblems that the
current subproblem depends on. It also *causes* the current
subproblem to depend only on (up to) two preceding subproblems, by
propagating best-solution information even through subproblems
ending at intervals that are not part of the optimal subset being
built.