Switch CodeQL to advanced and set up Python 3.11
CodeQL currently defaults to Python 3.10, including in the basic
configuration (which doesn't use/require a workflow file in the
repository). I was using that before, but it doesn't successfully
install the dependencies now, probably because of the switch from
Python 3.10 to Python 3.11 that his project has just underwent.

This uses an advanced configuration instead, with an added step to
run the setup-python action, specifying version 3.11.